requirements_modal.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
faiss-gpu-cu12>=1.8.0  # Совместим с CUDA 12.8[3]
fastparquet>=2024.2.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
gradio>=4.0.0          # Для веб-интерфейса
grpclib==0.4.7
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
modal==1.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numpy==1.26.4  # Специфичная версия, совместимая с torch
openai>=1.0.0          # Для работы с Nebius API[6]
packaging==25.0
pandas==2.2.2
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
pyarrow>=14.0.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==2.7.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
spacy[cuda-autodetect]==3.7.5
sympy==1.14.0
synchronicity==0.9.13
textacy==0.13.0     # Для извлечения ключевых слов
textblob==0.17.1    # Для анализа тональности
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.2.2
tqdm==4.66.4
transformers==4.52.4
typer>=0.9
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchfiles==1.0.5
yarl==1.20.0
# Для загрузки модели с помощью transformers
accelerate>=0.21.0
# Для 8-битной и 4-битной квантизации моделей
bitsandbytes>=0.43.0
# Для работы с pipeline и генерацией текста
sentencepiece
# LlamaIndex пакеты
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
